requirements.txt
fastapi
uvicorn
# easyocr
pillow
numpy
pytesseract
imutils
